+++ src/schema/payload/typeDefs.ts
@@ -36,7 +36,7 @@ const typeDefs = gql`
     reference_system: String
     regime: String
     longitude: Float
-    lifespan_years: Int
+    lifespan_years: Float
     epoch: Date
     mean_motion: Float
     raan: Float
@@ -67,7 +67,7 @@ const typeDefs = gql`
     apoapsis_km: Float
     inclination_deg: Float
     period_min: Float
-    lifespan_years: Int
+    lifespan_years: Float
     epoch: Date
     mean_motion: Float
     raan: Float
